@@ -101,15 +101,12 @@
     "chance": "0.8.0",
     "cheerio": "0.19.0",
     "coveralls": "2.11.16",
-    "eslint": "3.19.0",
-    "eslint-config-opencollective": "1.1.5",
-    "eslint-config-standard": "10.2.1",
-    "eslint-plugin-import": "1.16.0",
-    "eslint-plugin-mocha": "4.5.0",
-    "eslint-plugin-node": "5.1.1",
-    "eslint-plugin-promise": "3.5.0",
-    "eslint-plugin-react": "6.2.0",
-    "eslint-plugin-standard": "3.0.1",
+    "eslint": "^4.19.1",
+    "eslint-config-opencollective": "^1.1.7",
+    "eslint-plugin-import": "^2.11.0",
+    "eslint-plugin-mocha": "^5.0.0",
+    "eslint-plugin-node": "^6.0.1",
+    "eslint-plugin-react": "^7.7.0",
     "graphiql": "0.8.0",
     "jsdoc": "3.5.5",
     "jsdoc-simple-theme": "1.2.6",
